fix: changing result variable for each case
This commit adds the value of fields to some of the queries, so when the result of a query is sent to the final user, it will not show the buffer and other info. In case that the buffer is not sent, the fields value breaks the code, so it was not added.